refactor: migrate project to Next.js with TypeScript
