+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "dev": "next dev"
+    "start": "next start",
+    "dev": "next dev",
+    "build": "next build"
   },
   "keywords": [],
   "author": "",
@@ -14,8 +15,6 @@
   "type": "module",
   "dependencies": {
     "axios": "^1.9.0",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
     "next": "^15.3.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
