--- pages/api/projects.js
@@ -1,5 +1,0 @@
-// pages/api/projects.js
-
-export default function handler(req, res) {
-  res.status(200).json({ message: 'Projects list' });
-}
