+++ sections/samples/get.ts
@@ -24,3 +24,4 @@ request.get(
         }
     }
 );
+export {}
